+++ server/server.go
@@ -2,12 +2,9 @@ package server
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"grafana-matrix-forwarder/cfg"
-	"grafana-matrix-forwarder/grafana"
 	"grafana-matrix-forwarder/matrix"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"time"
@@ -18,6 +15,7 @@ type Server struct {
 	ctx               context.Context
 	matrixWriteCloser matrix.WriteCloser
 	appSettings       cfg.AppSettings
+	metrics           serverMetrics
 }
 
 // BuildServer builds a Server instance based on the provided context.Context, a matrix.WriteCloser, and the cfg.AppSettings
@@ -26,6 +24,7 @@ func BuildServer(ctx context.Context, matrixWriteCloser matrix.WriteCloser, appS
 		ctx:               ctx,
 		matrixWriteCloser: matrixWriteCloser,
 		appSettings:       appSettings,
+		metrics:           serverMetrics{},
 	}
 }
 
@@ -36,6 +35,19 @@ func (server Server) Start() (err error) {
 	mux.Handle("/api/v0/forward", http.HandlerFunc(
 		func(response http.ResponseWriter, request *http.Request) {
 			err = server.handleGrafanaAlert(response, request)
+			server.metrics.totalForwardCount++
+			if err != nil {
+				server.metrics.failForwardCount++
+				log.Print(err)
+				response.WriteHeader(500)
+			} else {
+				server.metrics.successForwardCount++
+			}
+		},
+	))
+	mux.Handle("/metrics", http.HandlerFunc(
+		func(response http.ResponseWriter, request *http.Request) {
+			err = server.handleMetricsRequest(response)
 			if err != nil {
 				log.Print(err)
 				response.WriteHeader(500)
@@ -78,58 +90,4 @@ func (server Server) Start() (err error) {
 		err = nil
 	}
 	return
-}
-
-func (server Server) handleGrafanaAlert(response http.ResponseWriter, request *http.Request) error {
-	bodyBytes, err := getRequestBodyAsBytes(request)
-	if err != nil {
-		return err
-	}
-	if server.appSettings.LogPayload {
-		logPayload(request, bodyBytes)
-	}
-
-	roomID, err := getRoomIDFromURL(request)
-	if err != nil {
-		return err
-	}
-
-	alert, err := getAlertPayloadFromRequestBody(bodyBytes)
-	if err != nil {
-		return err
-	}
-
-	log.Printf("alert received (%s) - forwarding to room: %s", alert.FullRuleID(), roomID)
-
-	err = grafana.ForwardAlert(server.matrixWriteCloser.GetWriter(), roomID, alert, server.appSettings.ResolveMode)
-	if err != nil {
-		return err
-	}
-
-	response.WriteHeader(200)
-	_, err = response.Write([]byte("OK"))
-	return err
-}
-
-func logPayload(request *http.Request, bodyBytes []byte) {
-	log.Printf("%s request received at URL: %s", request.Method, request.URL.String())
-	body := string(bodyBytes)
-	fmt.Println(body)
-}
-
-func getRoomIDFromURL(request *http.Request) (string, error) {
-	roomIds, ok := request.URL.Query()["roomId"]
-	if !ok || len(roomIds[0]) < 1 {
-		return "", fmt.Errorf("url param 'roomId' is missing")
-	}
-	return roomIds[0], nil
-}
-
-func getAlertPayloadFromRequestBody(bodyBytes []byte) (alertPayload grafana.AlertPayload, err error) {
-	err = json.Unmarshal(bodyBytes, &alertPayload)
-	return
-}
-
-func getRequestBodyAsBytes(request *http.Request) ([]byte, error) {
-	return ioutil.ReadAll(request.Body)
 }