+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 ## [Unreleased]
 
 ### Features
+- (9ef976e) feat: export metrics on alert state ([#17](https://gitlab.com/hectorjsmith/grafana-matrix-forwarder/issues/17))
+- (b86caa0) feat: export forward count metrics ([#17](https://gitlab.com/hectorjsmith/grafana-matrix-forwarder/issues/17))
+- (078cace) feat: add metrics endpoint ([#17](https://gitlab.com/hectorjsmith/grafana-matrix-forwarder/issues/17))
 - (0399f02) feat: support resolving alerts with reactions ([#11](https://gitlab.com/hectorjsmith/grafana-matrix-forwarder/issues/11))
 - (5cf2fdb) feat: load id fields from alert payload ([#11](https://gitlab.com/hectorjsmith/grafana-matrix-forwarder/issues/11))
 - (aa30c24) feat: support for sending reactions ([#11](https://gitlab.com/hectorjsmith/grafana-matrix-forwarder/issues/11))